+++ http/exchange_update_test.go
@@ -106,7 +106,7 @@ func TestHTTPUpdateExchange(t *testing.T) {
 	}
 
 	testServer := HTTPServer{
-		app:         data.NewReserveData(nil, nil, nil, nil, nil, nil, setting),
+		app:         data.NewReserveData(nil, nil, nil, nil, nil, nil, nil, setting),
 		core:        core.NewReserveCore(nil, nil, setting),
 		metric:      testStorage,
 		authEnabled: false,